Slowly introducing the Setup frame marshalling code
Much of the code written tentatively, still very much trying to find Go
best practices - notably I'm uncomfortable about how I'm using primitive
types here, but happy to experiment with eg. uint16 for the Flags
variable; if I can get type safety for that kind of thing, that's really
nice.

Most of the approach to how this marshalling code flows is grabbed
straight from the Java impl, not sure about how we're leveraging
packages to get the setup.<FieldName>(*codec.Form) signature functions -
it seems it lends the same zero-allocation as the pattern does in Java,
but I'm not sure it's the best way to get that effect.. Something tells
me there's a smarter way to do this in Go, not sure how yet.

Rectified a bug in the Java impl that may warrant a patch; if you tell
the Java impl to create a Setup package with the `metadata` flag high,
but with zero-length metdata, it'll generate an invalid packet.

Overall, tests are now failing where I'd hoped for them to, so tomorrow
I should be able to TDD my way forward.