Make frame length field not be a concern of the Frame codec
The frame length field is not part of the regular Frame data, it's
specific to TCP and other protocols that do not provide framing for us.

Hence, it should not live in the codec, but with the encoder/decoder
that we use for TCP.

Which, yes, that means those encoder/decoder structs need to be moved
to a transport-specific package.