Spiking out TCP server via TCK
It's, obviously, a giant mess - but I like keeping it all in this big
file for now as I hash out APIs and wiring. Intention is to try and get
the server to pass most (all?) of the TCK, using the java client as the
client impl, and then refactor and do the same for the client side.