Breakout time!
Move all the spiky stuff out of the TCK package. This may
be the first of several reorgs, but I feel good about this
first move.

Effectively, cordon off an API section with public interfaces under `rs`
(for "ReactiveSocket", easily confused with "ReactiveStream", but I
don't think that matters much), add a public `transport` package that
builds on `rs` to provide specific transport implementations, and
finally move everything else into an `internal` package.

That internal package is now structured as a set of russian dolls:

- codec, implements the binary encoding
- frame, implements the Frame memory abstraction and read/encode API
- proto, implements the semantics of the protocol

There's one clear drawback here, which is that it is impossible for
third party libraries to pull in these packages and implement their own
transports.

Effectively, we need an SPI package of some kind. Leaving that as a
future excercise though, YAGNI, KISS et cetera.