Protect Send with a mutex
I really hate this, for several reasons. Partially because locks in a
place like this is a scourge, if there is contention then every writer
will be ready to write at any second, they shouldn't need to wait for OS
signals.

But more importantly - it makes it impossible for the implementor of
Send to batch messages, which will destroy network performance since it
has no way to tell when to flush, it'll have to flush for each message.

This whould be replaced with a good ring buffer impl instead, that would
be ideal for this.