Add assertions to ensure RequestHandler#HandleChannel behaves ok
We may want to actually invert this later, and allow this behavior,
but I cannot see a legitimate use case for it right now, and it would
be a good bit of complexity to work around (eg. queuing inbound requests
until the appropriate behavior is seen, and dealing with timeouts etc).

If we ever come across a legitimate use case, we can revisit this.