Make Subscriber specific to Payload
I feel extremely uncomfortable about this; at the same time, it felt
like I was doing the wrong thing entirely trying to adhere to a
bastardized API designed for a language with generics; better make the
interface fit my needs.

This is not the end of this story, I suspect; before a stable release,
the overall design pattern needs to be revisited. There must be a
sensible way to design a reactive stream API that fits in with Go,
this one really does not seem like it does :/