Remove rs.Payload from channel initializer API
Not sure at all this is a sensible idea - it makes it really tricky
to route channels.. But the protocol is clearly designed this way,
and this is how the Java API looks. Prefer conformism here, since I'm
not the one designing the protocol.

Likewise, add handling for that initial payload getting treated as a
message bound for the Application Subscriber.