Handle EOF, make TCK a bit non-deterministic
Got stuck on TCK for a bit, until I realized the Java impl uses a
separate thread to handle 'marbles' from that which handles calls to
`request` and so on; the tests depend on this, and assume that it will
always take longer to start this new thread than to run the next script
command.

Don't really like this, since it makes the tests non-deterministic, but
it doesn't affect the actual impl code and it makes the tests pass
pending feedback on how to proceed.