Remove `stream` struct from proto
Protocol impl is still really hard to follow in big aspects, but this
helps a bit. Largely, it highlights a lifecycle realization:

    Channels are made up of two independent streams.

That is, I had previously assumed the two streams in a channel had to
both exist, or both be closed - that seems to be wrong, they exist
independently. Hence, when the server closes it's subscription to the
remote, inbounc, stream in a channel, the local, outbound, stream can
remain open.

Hence, the subscription to the local stream in a channel can be treated
as the subscription in a regular single-plex stream.

Also, this commit tries to highlight that there are for main reactive
components involved for each StreamID:

    Local, outbound, stream:
      Subscription
      Subscriber
    Remote, inbound, stream:
      Subscription
      Subscriber

Out of these, two are not currently needed to be referenced by the
Protocol (the ones our local application makes calls to), and two are
(the ones the remote transport is sending messages to):

    Recieves messages from transport:
      Remote stream subscriber
      Local stream subscription

    Receives application calls:
      Local stream subscriber (and forwards TO transport)
      Remote stream subscription (and forwards TO transport)

All in all, there's a lot of concepts to juggle at the same time here. I
suspect the right next step would be to find an abstraction that cuts
the number of things to juggle mentally at the same time down to 2.